refact(WordListView): 1. 뷰와 뷰모델 분리(코어데이터 관련 코드 분리) 2. 불필요한 코어데이터 관련 코드 제거 3. 주석 추가(설명) Close #1
